gitlab/requirements.txt
python-gitlab==1.7.0
GitPython==2.1.11